Move codec test helper to an internal package
