+++ codec/codec.go
@@ -1,3 +1,5 @@
+// Package codec contains sub-packages with different codecs that can be used
+// to encode and decode entities in Storm.
 package codec
 
 // EncodeDecoder represents a codec used to encode and decode entities.
